+++ app/src/main/java/com/mkulesh/micromath/widgets/CustomTextView.java
@@ -210,7 +210,7 @@ public class CustomTextView extends AppCompatTextView implements OnLongClickList
     @Override
     public int getBaseline()
     {
-        return (int) ((this.getMeasuredHeight() - getPaddingBottom() + getPaddingTop()) / 2);
+        return ((this.getMeasuredHeight() - getPaddingBottom() + getPaddingTop()) / 2);
     }
 
     @Override
